third_party/blink/web_tests/external/wpt/focus/activeelement-after-focusing-different-site-iframe-expected.txt

This is a testharness.js-based test.
[FAIL] Check result
  assert_equals: Check log expected "outeronload,activeElement:BODY,willfocusiframe,didfocusiframe,activeElement:IFRAME,willbluriframe,didbluriframe,activeElement:BODY,willspineventloop,innerbodyfocus,innerbodyblur," but got "outeronload,activeElement:BODY,willfocusiframe,didfocusiframe,activeElement:IFRAME,willbluriframe,didbluriframe,activeElement:BODY,willspineventloop,innerbodyblur,"
Harness: the test ran to completion.
